% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_fit.R
\name{coef,market_fit-method}
\alias{coef,market_fit-method}
\title{Estimated coefficients of a fitted market model.}
\usage{
\S4method{coef}{market_fit}(object)
}
\arguments{
\item{object}{A fitted model object.}
}
\value{
A vector of estimated model coefficients.
}
\description{
Returns the coefficients of the fitted model.
}
\examples{
\donttest{
# initialize the model using the houses dataset
model <- new(
  "diseq_deterministic_adjustment", # model type
  subject = ID, time = TREND, quantity = HS, price = RM,
  demand = RM + TREND + W + CSHS + L1RM + L2RM + MONTH,
  supply = RM + TREND + W + L1RM + MA6DSF + MA3DHF + MONTH,
  fair_houses(), # data
  correlated_shocks = FALSE # allow shocks to be correlated
)

# estimate the model.
est <- estimate(model, control = list(maxit = 1e+5), method = "BFGS")

# access the estimated coefficients
coef(est)
}
}
