% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_basic.R
\docType{class}
\name{diseq_basic-class}
\alias{diseq_basic-class}
\title{Basic disequilibrium model with unknown sample separation.}
\description{
The basic disequilibrium model consists of three equations. Two of them are
the demand and supply equations. In addition, the model replaces the market clearing condition
with the short side rule. The model is estimated using full information maximum likelihood.

\deqn{D_{nt} = X_{d,nt}'\beta_{d} + u_{d,nt},}
\deqn{S_{nt} = X_{s,nt}'\beta_{s} + u_{s,nt},}
\deqn{Q_{nt} = \min\{D_{nt},S_{nt}\}.}
}
