% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_deterministic_adjustment.R
\docType{class}
\name{diseq_deterministic_adjustment-class}
\alias{diseq_deterministic_adjustment-class}
\title{Disequilibrium model with deterministic price dynamics.}
\description{
Disequilibrium model with deterministic price dynamics.
}
