% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_stochastic_adjustment.R
\docType{class}
\name{diseq_stochastic_adjustment-class}
\alias{diseq_stochastic_adjustment-class}
\title{Disequilibrium model with stochastic price dynamics.}
\description{
Disequilibrium model with stochastic price dynamics.
}
