% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_fiml.R
\docType{class}
\name{eq_fiml-class}
\alias{eq_fiml-class}
\title{Equilibrium model estimated using full-information maximum likelihood.}
\description{
Equilibrium model estimated using full-information maximum likelihood.
}
