% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_deterministic_adjustment.R
\docType{class}
\name{equation_deterministic_adjustment-class}
\alias{equation_deterministic_adjustment-class}
\title{Deterministic adjustment disequilibrium model equation class}
\description{
Deterministic adjustment disequilibrium model equation class
}
