% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_directional.R
\docType{class}
\name{equation_directional-class}
\alias{equation_directional-class}
\title{Directional disequilibrium model equation class}
\description{
Directional disequilibrium model equation class
}
