% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_stochastic_adjustment.R
\docType{class}
\name{equation_stochastic_adjustment-class}
\alias{equation_stochastic_adjustment-class}
\title{Stochastic adjustment disequilibrium model equation class}
\description{
Stochastic adjustment disequilibrium model equation class
}
