% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{get_aggregate_demand}
\alias{get_aggregate_demand}
\alias{get_aggregate_demand,model_base-method}
\title{Demand aggregation.}
\usage{
get_aggregate_demand(object, estimation)

\S4method{get_aggregate_demand}{model_base}(object, estimation)
}
\arguments{
\item{object}{A model object.}

\item{estimation}{A model estimation object (i.e. a \code{\link[bbmle]{mle2}} object).}
}
\value{
The sum of the estimated demanded quantities.
}
\description{
Sets the model's parameters given in the estimation object and calculates the
sample's aggregate demand.
}
\seealso{
get_estimated_demanded_quantities
}
