% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{get_aggregate_supply}
\alias{get_aggregate_supply}
\alias{get_aggregate_supply,model_base-method}
\title{Supply aggregation.}
\usage{
get_aggregate_supply(object, estimation)

\S4method{get_aggregate_supply}{model_base}(object, estimation)
}
\arguments{
\item{object}{A model object.}

\item{estimation}{A model estimation object (i.e. a \code{\link[bbmle]{mle2}} object).}
}
\value{
The sum of the estimated supplied quantities.
}
\description{
Sets the model's parameters given in the estimation object and calculates sample's
aggregate supply.
}
\seealso{
get_estimated_supplied_quantities
}
