% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_base.R
\name{get_correlation_variable}
\alias{get_correlation_variable}
\alias{get_correlation_variable,system_base-method}
\title{Correlation variable name.}
\usage{
get_correlation_variable(object)

\S4method{get_correlation_variable}{system_base}(object)
}
\arguments{
\item{object}{A system object.}
}
\description{
Returns the variable name for the correlation coefficient parameter of the model.
}
