% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_base.R
\name{get_estimated_shortage_indices}
\alias{get_estimated_shortage_indices}
\alias{get_estimated_shortage_indices,diseq_base-method}
\title{Indices of estimated shortages.}
\usage{
get_estimated_shortage_indices(object, estimation)

\S4method{get_estimated_shortage_indices}{diseq_base}(object, estimation)
}
\arguments{
\item{object}{A disequilibrium model object.}

\item{estimation}{A model estimation object (i.e. a \code{\link[bbmle]{mle2}} object).}
}
\description{
Returns the indices for which the estimated shortages of the market are non-negative.
}
