% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{get_estimated_supplied_quantities}
\alias{get_estimated_supplied_quantities}
\alias{get_estimated_supplied_quantities,model_base-method}
\title{Estimated supplied quantities.}
\usage{
get_estimated_supplied_quantities(object, estimation)

\S4method{get_estimated_supplied_quantities}{model_base}(object, estimation)
}
\arguments{
\item{object}{A model object.}

\item{estimation}{A model estimation object (i.e. a \code{\link[bbmle]{mle2}} object).}
}
\value{
A vector with the estimated supplied quantities.
}
\description{
Sets the model's parameters given in the estimation object and calculates the estimated
supplied quantity for each observation.
}
\seealso{
get_aggregate_supply
}
