% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_base.R
\name{get_marginal_effect_at_mean}
\alias{get_marginal_effect_at_mean}
\alias{get_marginal_effect_at_mean,diseq_base-method}
\title{Marginal effects at the mean.}
\usage{
get_marginal_effect_at_mean(object, estimation, variable)

\S4method{get_marginal_effect_at_mean}{diseq_base}(object, estimation, variable)
}
\arguments{
\item{object}{A disequilibrium model object.}

\item{estimation}{A model estimation object (i.e. a \code{\link[bbmle]{mle2}} object).}

\item{variable}{Variable name for which the effect is calculated.}
}
\description{
Returns the estimated marginal effects evaluated at the mean.
}
