% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_base.R
\name{get_mean_marginal_effect}
\alias{get_mean_marginal_effect}
\alias{get_mean_marginal_effect,diseq_base-method}
\title{Mean marginal effects}
\usage{
get_mean_marginal_effect(object, estimation, variable)

\S4method{get_mean_marginal_effect}{diseq_base}(object, estimation, variable)
}
\arguments{
\item{object}{A disequilibrium model object.}

\item{estimation}{A model estimation object (i.e. a \code{\link[bbmle]{mle2}} object).}

\item{variable}{Variable name for which the effect is calculated.}
}
\description{
Returns the average estimated marginal effect.
}
