% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{get_model_description}
\alias{get_model_description}
\alias{get_model_description,model_base-method}
\title{Model description.}
\usage{
get_model_description(object)

\S4method{get_model_description}{model_base}(object)
}
\arguments{
\item{object}{A model object.}
}
\value{
A string representation of the model.
}
\description{
A unique identifying string for the model.
}
