% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_base.R
\name{get_normalized_shortages}
\alias{get_normalized_shortages}
\alias{get_normalized_shortages,diseq_base-method}
\title{Normalized shortages.}
\usage{
get_normalized_shortages(object, parameters)

\S4method{get_normalized_shortages}{diseq_base}(object, parameters)
}
\arguments{
\item{object}{A disequilibrium model object.}

\item{parameters}{A vector of parameters at which the shortages are evaluated.}
}
\description{
Returns the estimated shortages normalized by the variance of the difference of the shocks.
}
