% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{get_number_of_observations}
\alias{get_number_of_observations}
\alias{get_number_of_observations,model_base-method}
\title{Number of observations.}
\usage{
get_number_of_observations(object)

\S4method{get_number_of_observations}{model_base}(object)
}
\arguments{
\item{object}{A model object.}
}
\value{
The number of used observations.
}
\description{
Returns the number of observations that are used by an initialized model. The number of
used observations may differ
from the numbers of observations of the data set that was passed to the model's initialization.
}
