% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_base.R
\name{get_prefixed_independent_variables}
\alias{get_prefixed_independent_variables}
\title{Independent variable names.}
\usage{
get_prefixed_independent_variables(object)
}
\arguments{
\item{object}{An equation object.}
}
\value{
A vector with the independent variable names.
}
\description{
The names of the independent variables are constructed by concatenating the equation prefix
with the column names of the data \code{tibble}.
}
