% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_base.R
\name{get_prefixed_price_variable}
\alias{get_prefixed_price_variable}
\alias{get_prefixed_price_variable,equation_base-method}
\title{Price coefficient variable name.}
\usage{
get_prefixed_price_variable(object)

\S4method{get_prefixed_price_variable}{equation_base}(object)
}
\arguments{
\item{object}{An equation object.}
}
\description{
Returns the variable name for the price coefficient parameter of the equation.
}
