% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_base.R
\name{get_relative_shortages}
\alias{get_relative_shortages}
\alias{get_relative_shortages,diseq_base-method}
\title{Relative shortages.}
\usage{
get_relative_shortages(object, parameters)

\S4method{get_relative_shortages}{diseq_base}(object, parameters)
}
\arguments{
\item{object}{A disequilibrium model object.}

\item{parameters}{A vector of parameters at which the shortages are evaluated.}
}
\value{
A vector with the estimated normalized shortages.
}
\description{
Returns the estimated shortages normalized by the estimated supplied quantity.
}
