% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_base.R
\name{get_shortage_probabilities}
\alias{get_shortage_probabilities}
\alias{get_shortage_probabilities,diseq_base-method}
\title{Shortage probabilities.}
\usage{
get_shortage_probabilities(object, parameters)

\S4method{get_shortage_probabilities}{diseq_base}(object, parameters)
}
\arguments{
\item{object}{A disequilibrium model object.}

\item{parameters}{A vector of parameters at which the shortage probabilities are evaluated.}
}
\description{
Returns the estimated shortage probabilities.
}
