% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{get_supply_descriptives}
\alias{get_supply_descriptives}
\alias{get_supply_descriptives,model_base-method}
\title{Supply descriptive statistics}
\usage{
get_supply_descriptives(object)

\S4method{get_supply_descriptives}{model_base}(object)
}
\arguments{
\item{object}{A model object.}
}
\value{
A data \code{tibble} containing descriptive statistics.
}
\description{
Calculates and returns basic descriptive statistics for the model's supply data. Factor
variables are excluded from  the calculations.
}
