% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R, R/diseq_basic.R,
%   R/diseq_deterministic_adjustment.R, R/diseq_directional.R,
%   R/diseq_stochastic_adjustment.R, R/equilibrium_model.R
\name{gradient}
\alias{gradient}
\alias{gradient,diseq_basic-method}
\alias{gradient,diseq_deterministic_adjustment-method}
\alias{gradient,diseq_directional-method}
\alias{gradient,diseq_stochastic_adjustment-method}
\alias{gradient,equilibrium_model-method}
\title{Gradient}
\usage{
gradient(object, parameters)

\S4method{gradient}{diseq_basic}(object, parameters)

\S4method{gradient}{diseq_deterministic_adjustment}(object, parameters)

\S4method{gradient}{diseq_directional}(object, parameters)

\S4method{gradient}{diseq_stochastic_adjustment}(object, parameters)

\S4method{gradient}{equilibrium_model}(object, parameters)
}
\arguments{
\item{object}{A model object.}

\item{parameters}{A vector of parameters at which the gradient is to be evaluated.}
}
\value{
The opposite of the model log likelihood's gradient.
}
\description{
Returns the gradient of the opposite of the log-likelihood evaluated at the passed
parameters.
}
