% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R, R/hessian_basic.R,
%   R/hessian_directional.R
\name{hessian}
\alias{hessian}
\alias{hessian,diseq_basic-method}
\alias{hessian,diseq_directional-method}
\title{Hessian}
\usage{
hessian(object, parameters)

\S4method{hessian}{diseq_basic}(object, parameters)

\S4method{hessian}{diseq_directional}(object, parameters)
}
\arguments{
\item{object}{A model object.}

\item{parameters}{A vector of parameters at which the hessian is to be evaluated.}
}
\value{
The opposite of the model log likelihood's hessian.
}
\description{
Returns the hessian of the opposite of the log-likelihood evaluated at the passed
parameters.
}
