% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R, R/diseq_base.R,
%   R/diseq_basic.R, R/diseq_deterministic_adjustment.R, R/diseq_directional.R,
%   R/diseq_stochastic_adjustment.R, R/eq_base.R, R/eq_2sls.R, R/eq_fiml.R
\name{initialize_model}
\alias{initialize_model}
\alias{initialize,diseq_base-method}
\alias{initialize,diseq_basic-method}
\alias{initialize,diseq_deterministic_adjustment-method}
\alias{initialize,diseq_directional-method}
\alias{initialize,diseq_stochastic_adjustment-method}
\alias{initialize,eq_base-method}
\alias{initialize,eq_2sls-method}
\alias{initialize,eq_fiml-method}
\title{Models' constructions}
\usage{
\S4method{initialize}{model_base}(
  .Object,
  model_type_string,
  verbose,
  key_columns,
  time_column,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  price_specification,
  use_correlated_shocks,
  data,
  system_initializer
)

\S4method{initialize}{diseq_base}(
  .Object,
  model_type_string,
  verbose,
  key_columns,
  time_column,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  price_specification,
  use_correlated_shocks,
  data,
  system_initializer
)

\S4method{initialize}{diseq_basic}(
  .Object,
  verbose = 0,
  key_columns,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  use_correlated_shocks = TRUE,
  data
)

\S4method{initialize}{diseq_deterministic_adjustment}(
  .Object,
  verbose = 0,
  key_columns,
  time_column,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  use_correlated_shocks = TRUE,
  data
)

\S4method{initialize}{diseq_directional}(
  .Object,
  verbose = 0,
  key_columns,
  time_column,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  use_correlated_shocks = TRUE,
  data
)

\S4method{initialize}{diseq_stochastic_adjustment}(
  .Object,
  verbose = 0,
  key_columns,
  time_column,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  price_specification,
  use_correlated_shocks = TRUE,
  data
)

\S4method{initialize}{eq_base}(
  .Object,
  model_type_string,
  verbose,
  key_columns,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  use_correlated_shocks,
  data,
  system_initializer
)

\S4method{initialize}{eq_2sls}(
  .Object,
  verbose = 0,
  key_columns,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  data
)

\S4method{initialize}{eq_fiml}(
  .Object,
  verbose = 0,
  key_columns,
  quantity_column,
  price_column,
  demand_specification,
  supply_specification,
  use_correlated_shocks = TRUE,
  data
)
}
\arguments{
\item{.Object}{The object to be Constructed.}

\item{model_type_string}{Model type string (internal).}

\item{verbose}{Verbosity level.}

\item{key_columns}{Key columns of the data set.}

\item{time_column}{The time column of the data set.}

\item{quantity_column}{The quantity variable of the data set.}

\item{price_column}{The price variable of the data set.}

\item{demand_specification}{A formula representation of the right hand side of the
demand equation.}

\item{supply_specification}{A formula representation of the right hand side of the
supply equation.}

\item{price_specification}{A formula representation of the price equation.}

\item{use_correlated_shocks}{Should the model be estimated using correlated shocks?}

\item{data}{The data set.}

\item{system_initializer}{System initializer (internal).}
}
\value{
The initialized model.
}
\description{
Models' constructions
}
\details{
## Common initialization

### Variable construction

  The constructor prepares the model's variables using the passed specifications. The
  specification strings are expected to follow the
  syntax of \code{\link[stats]{formula}}. The construction of the model's data uses the variables
  that are extracted by these specification. The demand variables are extracted by a formula that
  uses the \code{quantity_column} on the left hand side and the \code{demand_specification} on
  the right hand side of the formula. The supply variables are constructed by the and the
  \code{quantity_column} and the  \code{supply_specification}. In the case of the
  \code{\linkS4class{diseq_stochastic_adjustment}} model, the price dynamics' variables are
  extracted using the \code{quantity_column} and the \code{price_specification}

### Data preparation

  1. If the passed data set contains rows with NA values, they are dropped. If the verbosity
    level allows warnings, a warning is emitted reporting how many rows were dropped.

  2. After dropping the rows, factor levels may be invalidated. If needed the constructor
    readjusts the factor variables by removing the unobserved levels. Factor indicators and
    interaction terms are automatically created.

  3. The primary column is constructed by pasting the values of the key_columns.

  4. In the case of the \code{\linkS4class{diseq_directional}},
    \code{\linkS4class{diseq_deterministic_adjustment}}, and
    the \code{\linkS4class{diseq_stochastic_adjustment}} models, a column with lagged prices
    is constructed. Since lagged prices are unavailable for the observation of the first
    time points, these observations are dropped. If the verbosity level allows the emission
    of information messages, the constructor prints the number of dropped observations.

  5. In the case of the \code{\linkS4class{diseq_directional}},
    and the \code{\linkS4class{diseq_stochastic_adjustment}} models, a column with price
    differences is created.
}
\section{Functions}{
\itemize{
\item \code{initialize,diseq_basic-method}: Basic disequilibrium model constructor.

\item \code{initialize,diseq_deterministic_adjustment-method}: Disequilibrium model with deterministic price
adjustment constructor

\item \code{initialize,diseq_directional-method}: Directional disequilibrium model constructor

\item \code{initialize,diseq_stochastic_adjustment-method}: Disequilibrium model with stochastic price adjustment
constructor

\item \code{initialize,eq_2sls-method}: Two stage least squares equilibrium model constructor.

\item \code{initialize,eq_fiml-method}: Two stage least squares equilibrium model constructor.
}}

