% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R
\docType{class}
\name{market_model-class}
\alias{market_model-class}
\title{Model base class}
\description{
Model base class
}
\section{Slots}{

\describe{
\item{\code{logger}}{Logger object.}

\item{\code{key_columns}}{Vector of column names that uniquely identify data records. For panel data
this vector should contain an entity and a time point identifier.}

\item{\code{time_column}}{Column name for the time point data.}

\item{\code{explanatory_columns}}{Vector of explanatory column names for all model's equations.}

\item{\code{data_columns}}{Vector of model's data column names. This is the union of the quantity,
price and explanatory columns.}

\item{\code{columns}}{Vector of primary key and data column names for all model's equations.}

\item{\code{model_tibble}}{Model data \code{tibble}.}

\item{\code{model_type_string}}{Model type string description.}

\item{\code{system}}{Model's system of equations.}
}}

