% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_eq_fiml.R
\name{maximize_log_likelihood}
\alias{maximize_log_likelihood}
\alias{maximize_log_likelihood,cpp_eq_fiml-method}
\title{Maximize the log-likelihood.}
\usage{
maximize_log_likelihood(
  object,
  start,
  step,
  objective_tolerance,
  gradient_tolerance
)

\S4method{maximize_log_likelihood}{cpp_eq_fiml}(
  object,
  start,
  step,
  objective_tolerance,
  gradient_tolerance
)
}
\arguments{
\item{object}{A model object.}

\item{start}{Initializing vector.}

\item{step}{Optimization step.}

\item{objective_tolerance}{Objective optimization tolerance.}

\item{gradient_tolerance}{Gradient optimization tolerance.}
}
\value{
A list with the optimization output.
}
\description{
Maximizes the log-likelihood using the
\href{https://www.gnu.org/software/gsl/doc/html/multimin.html}{GSL} implementation of
the BFGS algorithm. This function is primarily intended for advanced usage. The
\code{\link{estimate}} functionality is a fast analysis-oriented alternative.
}
\seealso{
estimate
}
