% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R, R/diseq_basic.R,
%   R/diseq_deterministic_adjustment.R, R/diseq_directional.R,
%   R/diseq_stochastic_adjustment.R, R/eq_fiml.R
\name{minus_log_likelihood}
\alias{minus_log_likelihood}
\alias{minus_log_likelihood,diseq_basic-method}
\alias{minus_log_likelihood,diseq_deterministic_adjustment-method}
\alias{minus_log_likelihood,diseq_directional-method}
\alias{minus_log_likelihood,diseq_stochastic_adjustment-method}
\alias{minus_log_likelihood,eq_fiml-method}
\title{Minus log-likelihood.}
\usage{
minus_log_likelihood(object, parameters)

\S4method{minus_log_likelihood}{diseq_basic}(object, parameters)

\S4method{minus_log_likelihood}{diseq_deterministic_adjustment}(object, parameters)

\S4method{minus_log_likelihood}{diseq_directional}(object, parameters)

\S4method{minus_log_likelihood}{diseq_stochastic_adjustment}(object, parameters)

\S4method{minus_log_likelihood}{eq_fiml}(object, parameters)
}
\arguments{
\item{object}{A model object.}

\item{parameters}{A vector of parameters at which the function is to be evaluated.}
}
\description{
Returns the opposite of the log-likelihood. The package uses the likelihoods expressions that
are derived in \href{Karapanagiotis (2020)}{http://dx.doi.org/10.2139/ssrn.3525622}.
}
