% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R
\name{model_description}
\alias{model_description}
\alias{model_description,market_model-method}
\title{Model description.}
\usage{
model_description(object)

\S4method{model_description}{market_model}(object)
}
\arguments{
\item{object}{A model object.}
}
\value{
A string representation of the model.
}
\description{
A unique identifying string for the model.
}
