% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logger.R
\docType{class}
\name{model_logger-class}
\alias{model_logger-class}
\title{Logger class}
\description{
Logger class
}
\section{Slots}{

\describe{
\item{\code{verbosity}}{Controls the intensity of output messages. Errors are always printed.
Other than this, a value of
\describe{
 \item{1}{prints warnings,}
 \item{2}{prints basic information,}
 \item{3}{prints verbose information and,}
 \item{4}{prints debug information.}
}}
}}

