% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R
\name{number_of_observations}
\alias{number_of_observations}
\alias{number_of_observations,market_model-method}
\title{Number of observations.}
\usage{
number_of_observations(object)

\S4method{number_of_observations}{market_model}(object)
}
\arguments{
\item{object}{A model object.}
}
\value{
The number of used observations.
}
\description{
Returns the number of observations that are used by an initialized model. The number
of used observations may differ from the numbers of observations of the data set
that was passed to the model's initialization.
}
