% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_base.R
\name{prefixed_control_variables}
\alias{prefixed_control_variables}
\title{Control variable names.}
\usage{
prefixed_control_variables(object)
}
\arguments{
\item{object}{An equation object.}
}
\value{
A vector with the control variable names.
}
\description{
The controls of the equation are the independent variables without the price variable.
Their names are constructed by concatenating the equation prefix with the name of
the price column.
}
