% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_base.R
\name{prefixed_variance_variable}
\alias{prefixed_variance_variable}
\alias{prefixed_variance_variable,equation_base-method}
\title{Variance variable name.}
\usage{
prefixed_variance_variable(object)

\S4method{prefixed_variance_variable}{equation_base}(object)
}
\arguments{
\item{object}{An equation object.}
}
\value{
The variable name for the variance of the shock of the equation.
}
\description{
The variance variables is constructed by concatenating the equation prefix with
"VARIANCE".
}
