% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_base.R
\name{simulate_model}
\alias{simulate_model}
\alias{simulate_model,ANY-method}
\title{Simulate equilibrium model.}
\usage{
simulate_model(
  model_string,
  nobs,
  tobs,
  alpha_d,
  beta_d0,
  beta_d,
  eta_d,
  alpha_s,
  beta_s0,
  beta_s,
  eta_s,
  gamma,
  beta_p0,
  beta_p,
  sigma_d = 1,
  sigma_s = 1,
  sigma_p = 1,
  rho_ds = 0,
  rho_dp = 0,
  rho_sp = 0,
  seed = NA,
  verbose = 0
)

\S4method{simulate_model}{ANY}(
  model_string,
  nobs,
  tobs,
  alpha_d,
  beta_d0,
  beta_d,
  eta_d,
  alpha_s,
  beta_s0,
  beta_s,
  eta_s,
  gamma,
  beta_p0,
  beta_p,
  sigma_d = 1,
  sigma_s = 1,
  sigma_p = 1,
  rho_ds = 0,
  rho_dp = 0,
  rho_sp = 0,
  seed = NA,
  verbose = 0
)
}
\arguments{
\item{model_string}{Equilibrium model type. It should be either 'eq_2sls' or 'eq_fiml'.}

\item{nobs}{Number of simulated entities.}

\item{tobs}{Number of simulated dates.}

\item{alpha_d}{Demand price coefficient.}

\item{beta_d0}{Demand constant coefficient.}

\item{beta_d}{Demand exclusive controls' coefficients.}

\item{eta_d}{Demand common controls' coefficients.}

\item{alpha_s}{Supply price coefficient.}

\item{beta_s0}{Supply constant coefficient.}

\item{beta_s}{Supply exclusive controls' coefficients.}

\item{eta_s}{Supply common controls' coefficients.}

\item{gamma}{Price adjustment coefficient.}

\item{beta_p0}{Price equation's constant coefficient.}

\item{beta_p}{Price equation's control coefficients.}

\item{sigma_d}{Demand shock's standard deviation.}

\item{sigma_s}{Supply shock's standard deviation.}

\item{sigma_p}{Price equation shock's standard deviation.}

\item{rho_ds}{Demand and supply shocks' correlation coefficient.}

\item{rho_dp}{Demand and price shocks' correlation coefficient.}

\item{rho_sp}{Supply and price shocks' correlation coefficient.}

\item{seed}{Pseudo number generator seed.}

\item{verbose}{Verbosity level.}
}
\description{
Returns an initialized equilibrium model with simulated data. The simulated observations of the controls are in drawn
from a uniform distribution on the interval [2, 5]. Passing an NA price coefficient results into an error. Price
observations are not simulated. Instead, they are calculated so that the market clears. If the simulations results
into negative demanded quantities, supplied quantities or prices, the execution halts with an error.
}
