% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulation.R
\name{simulate_model_data}
\alias{simulate_model_data}
\alias{simulate_model_data,ANY-method}
\title{Simulate model data.}
\usage{
simulate_model_data(
  model_string,
  nobs,
  tobs,
  alpha_d,
  beta_d0,
  beta_d,
  eta_d,
  alpha_s,
  beta_s0,
  beta_s,
  eta_s,
  gamma,
  beta_p0,
  beta_p,
  sigma_d = 1,
  sigma_s = 1,
  sigma_p = 1,
  rho_ds = 0,
  rho_dp = 0,
  rho_sp = 0,
  seed = NA_integer_,
  price_generator = function(nobs) stats::rnorm(n = nobs, mean = 2.5, sd = 0.5),
  control_generator = function(nobs) stats::rnorm(n = nobs, mean = 2.5, sd = 0.5)
)

\S4method{simulate_model_data}{ANY}(
  model_string,
  nobs,
  tobs,
  alpha_d,
  beta_d0,
  beta_d,
  eta_d,
  alpha_s,
  beta_s0,
  beta_s,
  eta_s,
  gamma,
  beta_p0,
  beta_p,
  sigma_d = 1,
  sigma_s = 1,
  sigma_p = 1,
  rho_ds = 0,
  rho_dp = 0,
  rho_sp = 0,
  seed = NA_integer_,
  price_generator = function(nobs) stats::rnorm(n = nobs, mean = 2.5, sd = 0.5),
  control_generator = function(nobs) stats::rnorm(n = nobs, mean = 2.5, sd = 0.5)
)
}
\arguments{
\item{model_string}{Equilibrium model type. It should be either \code{eq_2sls} or \code{eq_fiml}.}

\item{nobs}{Number of simulated entities.}

\item{tobs}{Number of simulated dates.}

\item{alpha_d}{Demand price coefficient.}

\item{beta_d0}{Demand constant coefficient.}

\item{beta_d}{Demand exclusive controls' coefficients.}

\item{eta_d}{Demand common controls' coefficients.}

\item{alpha_s}{Supply price coefficient.}

\item{beta_s0}{Supply constant coefficient.}

\item{beta_s}{Supply exclusive controls' coefficients.}

\item{eta_s}{Supply common controls' coefficients.}

\item{gamma}{Price equation's stability factor.}

\item{beta_p0}{Price equation's constant coefficient.}

\item{beta_p}{Price equation's control coefficients.}

\item{sigma_d}{Demand shock's standard deviation.}

\item{sigma_s}{Supply shock's standard deviation.}

\item{sigma_p}{Price equation shock's standard deviation.}

\item{rho_ds}{Demand and supply shocks' correlation coefficient.}

\item{rho_dp}{Demand and price shocks' correlation coefficient.}

\item{rho_sp}{Supply and price shocks' correlation coefficient.}

\item{seed}{Pseudo number generator seed.}

\item{price_generator}{Pseudo number generator for prices.}

\item{control_generator}{Pseudo number generator for non-prices controls.}
}
\description{
Returns a data \code{tibble} with simulated data from a generating process that matches the
passed model string. The simulated observations of the controls are drawn from a uniform
distribution on the interval [2, 5].
}
