% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R
\name{supply_descriptives}
\alias{supply_descriptives}
\alias{supply_descriptives,market_model-method}
\title{Supply descriptive statistics}
\usage{
supply_descriptives(object)

\S4method{supply_descriptives}{market_model}(object)
}
\arguments{
\item{object}{A model object.}
}
\value{
A data \code{tibble} containing descriptive statistics.
}
\description{
Calculates and returns basic descriptive statistics for the model's supply data. Factor
variables are excluded from  the calculations.
}
