% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_base.R
\name{variable_names}
\alias{variable_names}
\alias{prefixed_const_variable}
\alias{prefixed_independent_variables}
\alias{prefixed_price_variable}
\alias{prefixed_control_variables}
\alias{prefixed_variance_variable}
\title{Variable name access}
\usage{
prefixed_const_variable(object)

prefixed_independent_variables(object)

prefixed_price_variable(object)

prefixed_control_variables(object)

prefixed_variance_variable(object)
}
\arguments{
\item{object}{An equation object.}
}
\value{
The prefixed variable name(s).
}
\description{
Methods that provide access to the prefixed variable names that the
package uses.

\code{prefixed_const_variable}: The constant coefficient name is
constructed by concatenating the equation prefix with \code{CONST}.

\code{prefixed_independent_variables}: The names of the independent
variables are constructed by concatenating the equation prefix with the column names
of the data \code{tibble}.

\code{prefixed_price_variable}: The price coefficient name is
constructed by concatenating the equation prefix with the name of the price column.

\code{prefixed_control_variables}: The controls of the equation are the
independent variables without the price variable. Their names are constructed by
concatenating the equation prefix with the name of the price column.

\code{prefixed_control_variables}: The variance variables is
constructed by concatenating the equation prefix with \code{VARIANCE}.
}
\section{Functions}{
\itemize{
\item \code{prefixed_const_variable}: Constant coefficient variable name.

\item \code{prefixed_independent_variables}: Independent variable names.

\item \code{prefixed_price_variable}: Price coefficient variable name.

\item \code{prefixed_control_variables}: Control variable names.

\item \code{prefixed_variance_variable}: Variance variable name.
}}

