% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_fit.R
\name{coef,market_fit-method}
\alias{coef,market_fit-method}
\title{Estimated coefficients of a fitted market model.}
\usage{
\S4method{coef}{market_fit}(object)
}
\arguments{
\item{object}{A fitted model object.}
}
\value{
A vector of estimated model coefficients.
}
\description{
Returns the coefficients of the fitted model.
}
\examples{
\donttest{
# estimate a model using the houses dataset
fit <- diseq_deterministic_adjustment(
  HS | RM | ID | TREND ~
  RM + TREND + W + CSHS + L1RM + L2RM + MONTH |
  RM + TREND + W + L1RM + MA6DSF + MA3DHF + MONTH,
  fair_houses(),  correlated_shocks = FALSE,
  estimation_options = list(control = list(maxit = 1e+6)))

# access the estimated coefficients
coef(fit)
}
}
