% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_basic.R
\docType{class}
\name{diseq_basic-class}
\alias{diseq_basic-class}
\title{Basic disequilibrium model with unknown sample separation.}
\description{
Basic disequilibrium model with unknown sample separation.
}
