% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseq_directional.R
\docType{class}
\name{diseq_directional-class}
\alias{diseq_directional-class}
\title{Directional disequilibrium model with sample separation.}
\description{
Directional disequilibrium model with sample separation.
}
