% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_fiml.R
\docType{class}
\name{eq_fiml-class}
\alias{eq_fiml-class}
\title{Equilibrium model estimated using full-information maximum likelihood.}
\description{
The equilibrium model consists of thee equations. The demand, the
supply and the market clearing equations. This model is estimated using full information
maximum likelihood.

\deqn{
  \begin{aligned}
  D_{nt} &= X_{d,nt}'\beta_{d} + P_{nt}\alpha_{d} + u_{d,nt}, \\
  S_{nt} &= X_{s,nt}'\beta_{s} + P_{nt}\alpha_{s} + u_{s,nt}, \\
  Q_{nt} &= D_{nt} = S_{nt} ,
  \end{aligned}
}
}
\seealso{
\code{\link{initialize_model}}
}
