% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_model.R
\name{get_demand_descriptives}
\alias{get_demand_descriptives}
\alias{get_demand_descriptives,market_model-method}
\title{Demand descriptive statistics}
\usage{
get_demand_descriptives(object)

\S4method{get_demand_descriptives}{market_model}(object)
}
\arguments{
\item{object}{A model object.}
}
\value{
A data \code{tibble} containing descriptive statistics.
}
\description{
Calculates and returns basic descriptive statistics for the model's demand data. Factor
variables are excluded from the calculations.
}
