% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_base.R
\name{get_prefixed_const_variable}
\alias{get_prefixed_const_variable}
\alias{get_prefixed_const_variable,equation_base-method}
\title{Constant coefficient variable name.}
\usage{
get_prefixed_const_variable(object)

\S4method{get_prefixed_const_variable}{equation_base}(object)
}
\arguments{
\item{object}{An equation object.}
}
\value{
The constant coefficient name.
}
\description{
The constant coefficient name is constructed by concatenating the equation prefix
with \code{CONST}.
}
